// src/components/Investments/Investments.types.ts
export type InvestmentSectionProps = {
	align?: string;
	bgColor?: string;
	children?: React.ReactNode;
	cols?: number;
	eyebrow?: string;
	grid?: boolean;
	icon?: string;
	iconBg?: string;
	maxWidth?: string;
	name?: string;
	title?: string;
}

export type StyledContentProps = {
	align?: string;
}

export type StyledIconProps = {
	align?: string;
	bgColor?: string;
}

export type StyledHeaderProps = {
	textAlign?: string;
}

export type StyledIconContainerProps = {
	align?: string;
}

export type StyledSectionProps = {
	align?: string;
	bgColor?: string;
}

export type StyledContainerProps = {
	align?: string;
	grid?: boolean;
	maxWidth?: string;
}

export type SimpleSectionProps = {
	align?: string;
	children?: React.ReactNode;
	grid?: boolean;
	name?: string;
	textColor?: string;
}

export type StyledParagraphProps = {
	align?: string;
	color: string;
	font?: string;
	maxWidth?: string;
	size?: string;
	weight?: string;
}